+++ R/globals.R
@@ -1,9 +1,10 @@
 utils::globalVariables(c(
-  ".category", ".colour", ".period", ".period_index", ".period_label",
+  ".absolute_change", ".category", ".colour", ".period", ".period_index",
+  ".period_label", ".transition",
   ".value", "category_left", "category_left_next", "category_right",
-  "category", "category_text", "category_x", "from", "label_from",
+  "category", "category_text", "category_x", "change_label", "from", "label_from",
   "label_to", "missing_from", "missing_to", "rank", "rank_change",
-  "rank_from", "rank_to", "rank_x", "status", "to", "transition_index",
+  "movement", "rank_from", "rank_to", "rank_x", "status", "to", "transition_index",
   "value_change", "value_from",
   "value_left", "value_right", "value_to", "value_x", "x", "y", "y_next"
 ))